requirements.txt
amqp==2.6.0
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.6
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
Django==3.0.8
django-cors-headers==3.4.0
djangorestframework==3.11.0
dnspython==1.16.0
future==0.18.2
gunicorn==20.0.4
honcho==1.0.1
idna==2.10
importlib-metadata==1.7.0
kombu==4.6.11
PyJWT==1.7.1
pymongo==3.10.1
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.9
vine==1.3.0
zipp==3.1.0
